moved navLinks.js to controllers as eventlistener.js
