+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/thuongb14/project-management-app#readme",
   "dependencies": {
+    "bcrypt": "^5.1.0",
     "body-parser": "^1.20.1",
     "express": "^4.18.2",
     "pg": "^8.8.0"
